Apply Alexandra's comments
