+++ src/plugins/intel_cpu/src/emitters/snippets/cpu_kernel_executor_table.hpp
@@ -13,16 +13,16 @@ namespace intel_cpu {
 template<typename Conf, typename KernelType>
 class CPUKernelExecutor : public snippets::KernelExecutor<Conf, KernelType> {
 public:
-     CPUKernelExecutor(ov::intel_cpu::MultiCachePtr kernel_kache, std::shared_ptr<Conf> c) :
-                       snippets::KernelExecutor<Conf, KernelType>(c), m_kernel_cache(std::move(kernel_kache)) {}
+     CPUKernelExecutor(ov::intel_cpu::MultiCachePtr kernel_cache, std::shared_ptr<Conf> c) :
+                       snippets::KernelExecutor<Conf, KernelType>(c), m_kernel_cache(std::move(kernel_cache)) {}
      struct Key {
-         explicit Key(std::shared_ptr<Conf> c) : config{c} {}
+         explicit Key(std::shared_ptr<Conf> c) : config{std::move(c)} {}
          const std::shared_ptr<Conf> config;
          size_t hash() const { return config->hash(); }
          bool operator==(const Key& rhs) const { return *config == *rhs.config; }
      };
     void update_kernel() override {
-        OPENVINO_ASSERT(m_config && m_config->is_complete(), "Update kernel was called with invalid config");
+        OPENVINO_ASSERT(m_config && m_config->is_completed(), "Update kernel was called with invalid config");
         OPENVINO_ASSERT(m_kernel_cache, "Invalid kernel cache pointer in CPUKernelExecutor::update_kernel()");
         const auto& lookup_result = m_kernel_cache->getOrCreate(Key(m_config),
                                                                 [this](const Key& k) {
@@ -37,7 +37,7 @@ protected:
     using snippets::KernelExecutor<Conf, KernelType>::m_config;
     using snippets::KernelExecutor<Conf, KernelType>::m_kernel;
     using snippets::KernelExecutor<Conf, KernelType>::compile_kernel;
-    /** CPU plugin cache impelemntation used to avoid redundant recompilations */
+    /** CPU plugin cache implementation is used to avoid redundant recompilations */
     ov::intel_cpu::MultiCachePtr m_kernel_cache;
 };
